requirements.txt
google-generativeai
python-dotenv
pyaudio
speechrecognition
pyttsx3
pygame
requests
dateparser
